backend/requirements.txt
flask
flask-socketio
flask-cors
pymongo
python-dotenv
eventlet
Werkzeug
requests
gunicorn
